Properties

+++ Monopoly/GradleProject/Monopoly/app/src/main/java/com/rob/monopoly/Interfaces/Improvable.java
@@ -8,8 +8,8 @@ public interface Improvable extends Ownable {
 
     boolean buildHouse();
     boolean demolishHouse();
-    int getHousePrice();
-    int getNumHouses();
+    double getHousePrice();
+    double getNumHouses();
     String getColourGroup();
     boolean ownsColourGroup();
 }
